requirements.txt
aiohttp==1.0.5
async-timeout==2.0.1
certifi==2018.4.16
chardet==3.0.4
discord.py==0.16.12
fuzzywuzzy==0.16.0
idna==2.7
multidict==4.3.1
python-Levenshtein==0.12.0
requests==2.19.1
urllib3==1.23
websockets==3.4
